refactor and restructure project
