+++ src/main/java/host_room/HostRoomInputBoundary.java
@@ -1,6 +1,6 @@
 package host_room;
 
-import leaveRoom.RoomModel;
+import models.RoomModel;
 
 public interface HostRoomInputBoundary {
     void hostRoom(RoomModel request);
