+++ src/main/java/leave_room/LeaveRoomInputBoundary.java
@@ -1,4 +1,4 @@
-package leaveRoom;
+package leave_room;
 
 public interface LeaveRoomInputBoundary {
     void leaveRoom(Integer roomId, Integer currUserId);
