+++ src/main/java/leave_room/LeaveRoomOutputBoundary.java
@@ -1,4 +1,4 @@
-package leaveRoom;
+package leave_room;
 
 public interface LeaveRoomOutputBoundary {
     void prepareHostOrJoinView(LeaveRoomResponseModel responseModel);
