+++ src/main/java/leave_room/LeaveRoomResponseModel.java
@@ -1,4 +1,4 @@
-package leaveRoom;
+package leave_room;
 
 public class LeaveRoomResponseModel {
     private boolean wasSuccess;
