+++ src/main/java/models/UserDBModel.java
@@ -1,14 +1,19 @@
-package signup_and_login;
+package models;
 
-public class UserModel {
+public class UserDBModel {
+    private final int userID;
     private final String username;
     private final String password;
 
-    public UserModel(String username, String password) {
+    public UserDBModel(int userID, String username, String password) {
+        this.userID = userID;
         this.username = username;
         this.password = password;
     }
 
+    public int getUserID() {
+        return this.userID;
+    }
     public String getUsername() {
         return this.username;
     }
