+++ src/main/java/signup_and_login/UserController.java
@@ -1,5 +1,7 @@
 package signup_and_login;
 
+import models.UserModel;
+
 public class UserController {
     private final UserInputBoundary inputBoundary;
