+++ src/main/java/signup_and_login/UserInputBoundary.java
@@ -1,5 +1,7 @@
 package signup_and_login;
 
+import models.UserModel;
+
 public interface UserInputBoundary {
     void handleUser(UserModel request);
 }
