+++ src/main/java/util/RemoteUtilities.java
@@ -1,8 +1,5 @@
 package util;
 
-import com.google.gson.JsonObject;
-import signup_and_login.UserDBModel;
-
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
